Use BIGINT primary key for clinical_event table (#11399)
* Use BIGINT primary key for clinical_event table

* Respond to PR feedback

* Update cgds files in src/test/resources in response to Gaofei's feedback

+++ src/main/java/org/cbioportal/legacy/model/ClinicalEvent.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 
 public class ClinicalEvent extends UniqueKeyBase {
     
-    private Integer clinicalEventId;
+    private Long clinicalEventId;
     @NotNull
     private String studyId;
     @NotNull
@@ -18,11 +18,11 @@ public class ClinicalEvent extends UniqueKeyBase {
     private Integer stopDate;
     private List<ClinicalEventData> attributes;
 
-    public Integer getClinicalEventId() {
+    public Long getClinicalEventId() {
         return clinicalEventId;
     }
 
-    public void setClinicalEventId(Integer clinicalEventId) {
+    public void setClinicalEventId(Long clinicalEventId) {
         this.clinicalEventId = clinicalEventId;
     }
